Telemetry: flattens list of lists when printing the summary for multi-page results.
This is a follow up from r188798.
On Android, one of the metrics (namely, surface_latencies), is a list per url,
and then was summarized as a list of lists.
Flatten it, so it can print the avg / stddev.

TEST=tools/perf/run_multipage_benchmarks -v --browser=android-content-shell --show-stdout scrolling_benchmark tools/perf/page_sets/top_25.json

BUG=
TBR=tonyg
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/12907007

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ca6cbdbd88661dfcff8718f9fc4246fb3d521daf