[Telemetry] Set all command line flags that performance_ui_tests sets.
This is the result of diffing the telemetry flags with the flags set during a
performance_ui_test run of the page cyclers. The hope is that some of these
will help with noise in the telemetry page cyclers.

BUG=196411
TEST=None
NOTRY=True

Review URL: https://codereview.chromium.org/12481020

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4be1c126fc44ce0b336e9daccc1e21911965bb08