[Telemetry] Add a WillRunPageSet()/DidRunPageSet() API.
This is useful for benchmarks to call immediately prior to and after
running a page_set. It will be used by the page cyclers to record memory and IO
statistics which are a delta of before and after the entire page set runs.

This involved adding the concept of perf results that are not associated with a
particular page.

BUG=None
TEST=tools/telemetry/run_tests --browser=system page_benchmark_results_unittest
NOTRY=True


Review URL: https://chromiumcodereview.appspot.com/13488005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8082f69facd7f879a95a0ae7d157c5b2dc763449