[Telemetry] Make page cyclers more telemetric.
This causes the page cyclers to use page_sets instead of just navigating to the
start page and letting the pages cycle themselves. This means we'll be using
Telemetry "properly" which will allow us to record new page sets to update the
page cyclers and it means that page cycler page sets are interchangeable with
other benchmarks.

The page cyclers still measure the same thing as they measured previously: the
time to load the page and perform a layout. However, they measure it differently
now. Instead of setting a new Date().getTime() in the cookie immediately before
navigating and then grabbing the finish Date in the onload handler, this now
just grabs performance.now() in the onload handler. This returns the number of
milliseconds since navigationStart where navigation start is the same as the
immediately before navigation time marked before. I've verified locally that the
times reported are consistent, and will monitor the bots to verify.

This involved moving the serving_dirs property up from the page to the page_set.

BUG=None
TEST=All page cyclers on linux and morejs on windows
NOTRY=True

Review URL: https://codereview.chromium.org/13817009

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 830356bc02ecd0a263d40240859983312d9ab0f5