Fix indexed_db page cycler.
This was broken when the test switched over to being run with telemetry because
telemetry waits for the document.readyState to be "complete" and a late document.write()
causes the readyState to go back to "loading".

This patch fixes that by changing the test's logging method to appending to a log
div rather than document.write()ing after the page has loaded.

BUG=235350
TEST=tools/perf/run_multipage_benchmarks --browser=system page_cycler tools/perf/page_sets/page_cycler/indexed_db/basic_insert.json

Review URL: https://codereview.chromium.org/14383008

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a54e8115696da36621cb06388cc9589270abc00e