[Telemetry] Fix scrolling benchmark on android.
The scrolling benchmark fails on android with a DidNotScrollException
on Calendar. This is because by default the viewport scales out to the
point where the page is not scrollable.

To work around this, we inject a meta tag which prevents the viewport
from being scaled out. This allows the page to be scrollable, even on
android. On desktop, this patch is a no-op.

BUG=None
TEST=tools/perf/run_multipage_benchmarks --browser=android-content-shell scrolling_benchmark tools/perf/page_sets/top_25.json --page-filter=calendar
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/14652009

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3516b83309f09866c626394e362833ca87e8eac1