Telemetry: Close connections after executing each action.
The reload benchmark is experiencing frequent timeouts which occur only when
running under Web Page Replay. This solution was suggested in crbug.com/240721
and it seems to fix the problem.

Also, some pages of the memory benchmark timeout flakily, and this is likely the
cause there, too. So it seems it's not enough to do the cleanup when doing
reloads, but it also needs to be done more frequently when a page runs multiple
actions.

BUG=240721
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/14806006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9afcdaf7c9ebbc3cb8b2ad55c0ad3d1b6b8afdc9