Generate and use LASTCHANGE for blink
The main reason for making this change is the fact that when pulling
Chromium and Blink code into Android, we have to make automatic changes
to the repository (for example, for generating *.mk files), and the Blink
commit we have as HEAD doesn't match any public commit in the main 
Blink repo. While via the LASTCHANGE file we could report a publicly 
visible revision.

Also, the way webkit_version.h header is generated now appears to be silly.
First, it doesn't leverage enough existing tools. Second, as it is noted in
webkit_user_agent.gypi, it is broken for incremental builds.

Review URL: https://chromiumcodereview.appspot.com/14973005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b0abfb679a836ccb7d6da8d6f3c9c942a9295c9a