Make mass-rename.py work on Unix.
Per 17.1.1.2 on http://docs.python.org/2/library/subprocess.html, passing
shell=True on Unix and having the command to run be specified as a sequence
means that "the first item specifies the command string, and any additional
items will be treated as additional arguments to the shell itself." This
results in incorrect behavior (and in particular, the popen.communicate() fails
with a return code of 1).

This CL switches to the recommended behavior of passing the command to run as a
string when specifying shell=True.

R=scottmg

Review URL: https://chromiumcodereview.appspot.com/16849002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ba9722fca8f50bbdcab984fd8834f63d23f1229a