Remove suppressions for bug 222601.
The warning was caused by race inside XFixes when running
ScreenCaptureDeviceTest and ScreenCapturerTest. ScreenCapturerDevice uses
XFixes on a parallel thread that later competes with ScreenCapturerTest.
The failure doesn't happen anymore because the tests have been moved to
different locations (content and webrtc), so the suppressions are not
needed anymore.

It's not completely clear what causes the failure in the first place.
XextFindDisplay() seems to be thread-safe (
http://cvsweb.xfree86.org/cvsweb/*checkout*/xc/lib/Xext/extutil.c?rev=1.8
) - it locks _Xglocal_lock before accessing global values.

BUG=222601
R=wez@chromium.org

Review URL: https://codereview.chromium.org/17412011

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 57d70683af99e601abb08f68651b873d74099475