[Telemetry] Detect if renderers are swapped out during perf profiling.
When a cross-origin navigation occurs, a new renderer process is used.
Since profiling is started prior to navigation, it is possible that the
interesting renderer for the new page is not profiled at all.

The best thing we can do in this case is to detect it and warn the user
to run with --single-process.

BUG=None
TEST=Perf profiler with typical_25 page cycler
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/16154037

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7280637a68bce7b521b0d43b89832db35dc85125