UMA histograms: add (temporarily) swap jank histograms
This adds 64 histograms which attempt to measure jank due to
paging.

We would like to run this for a short time for a Finch experiment,
to help understand the cost of moving to 64/64 on 2 GB chromebooks.
After we understand the data, we should be able to do away with
most of them.

The large number of histograms results from the cross product of
different situations which we want to isolate.  The structure of
each histogram is:

Platform.SwapJank.<event>.<measure>.<swap group>.<time interval>

where

<event> is TabSwitch or Scroll
<measure> is Cpu or Faults
<swap group> is Swap0, ... Swap3
<time interval> is Time1, ... Time4

When an event occurs (tab switch or scroll), we post tasks to the blocking
pool to read some measures at a few time intervals following the event
(100, 300, 1000 and 3000 milliseconds).  Each measure is an average
for the interval.  Measures are CPU utilization (%) and number of page
faults per second.  We further classify the event by the current
utilization of the swap device: up to 10%, 10-30%, 30-60%, >60%.

Also see https://codereview.chromium.org/17577003/.

BUG=252772
TEST=none

Review URL: https://chromiumcodereview.appspot.com/18231005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 64ff20f8ba015ca544303f9e32884797d72b99fd