Add a simple tool which enables automated bisecting of codebase with manual test
Creates a tool which enables bisection the codebase, with the user prompted to
manually test a given revision and enter whether it is good or bad. Useful in
tracking down the CL which caused a bug which can be triggered manually.

BUG=

Review URL: https://chromiumcodereview.appspot.com/18316008

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 19cbf13dc344795b0cd81169a26a5fcb3f885b04