Fix a serious memory leak in forwarder2.
The memory used for a thread cannot be reclaimed until pthread_join or
pthread_detach is called because the OS doesn't know whether
pthread_join might be called in the future. Since we never Join
Forwarders, we have to Detach them in order for the memory to be freed.

Also fix a very minor leak where we forgot to free the memory allocated
by getaddrinfo.

This was using so much virtual memory that it was causing the
host_forwarder process to be killed which caused the page cyclers on
android to load blank pages and appear really fast.

PERF SHERIFFS: This may make the android page cyclers appear to regress.
It is not a real regression, just actually loading the pages properly
now.

BUG=256897

Review URL: https://chromiumcodereview.appspot.com/18894004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f7a0c2a02286b80e3e3c9b4d02d497ff9f4caa64