Allow specifying scroll starting position
The espn page in top_25.json does not scroll because there is
flash in the scroll starting position (the middle point of the
page) that consumes the touch events.

This CL allows specifying the relative scroll starting position
to the webpage and change the scroll starting position for the
espn page to not fall into the flash area.

BUG=246385
TEST=the espn page in top_25.json pass on Pixel.

Review URL: https://chromiumcodereview.appspot.com/18182017

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a4d338aa5349f7944609c6e47c213c50aa7a5b64