Reland r212359: "Delete DeviceController instances before ..."
This CL was reverted speculatively but turned out not to be the culprit.

While the Forwarder class on the device could get deleted when it detected an
error on a socket, the corresponding DeviceController could stick around. This
resulted in an excessive amount of file descriptors staying open that could
make the net_unittests fail.

This CL adds an explicit unmap command that gets sent by the host to the device
right before the host controller gets deleted. When the device receives this
message, it deletes the DeviceController instance corresponding to the port
being unmapped.

BUG=242846
TBR=bulach@chromium.org

Review URL: https://codereview.chromium.org/19790002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 035fe5529ca007d1c45255a3bc71e8ae18c0c374