Adding extended report to memdump with resident pages bit reporting.
This change introduces a new switch (-x) which causes memdump to
add, to each mapping printed out, a bitmap containing resident bit
information for the mapping.
This extended report can be processed by the memsymbols.py script,
part of this CL. The aim of memsymbols.py is to extract the list of
symbols, for a given library mapped in the analyzed process, which was
resident in memory at a given point in time (when the memdump snapshot
was taken).

NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/19466005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8bde6bc3404ca01d1413490fb2948fa9dfe874ec