Revert 214498 "Add UMA histograms for home page, start page, and..."
We already collect at least some of this data, and I'm worried the changes here
will corrupt all our existing data.  Reverting ASAP to minimize the chance of
damage until we figure out the right way to do this.

Even if this turns out not to be corrupting data, it's still almost certainly
the wrong implementation (we should be building on all the existing framework
that already does this, including for non-prepopulated engines).

> Add UMA histograms for home page, start page, and default search engine
> hosts, as they are set in the Preferences file.
> 
> This adds histogram enum values for non-prepopulated DSE hosts. The expanded enum is used to log Preferences settings for home page, first start page, and DSE hosts.
> 
> TBR=battre@chromium.org, meacer@chromium.org, mpearson@chromium.org, rockot@chromium.org
> 
> BUG=263234
> 
> Review URL: https://codereview.chromium.org/20012003

TBR=bbudge@chromium.org

Review URL: https://codereview.chromium.org/21672002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 38a79e25b234f6125096a4f32a88c77eb71aa626