Add SimpleCache.HeaderSize histograms
We are planning to combine headers and body into one file, and are
planning to put the headers first, and want to know how much slack
space we will need to leave to minimize the chance we will need to move
the body (which is terribly expensive) while not wasting as much space
as we're saving in cluster loss by combining the files.

This change adds some histograms that log header size and size changes.

BUG=

Review URL: https://chromiumcodereview.appspot.com/21383002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a5fb24fa37e0498ed0025de4a08e0a67f50f13a4