Hook up better describing, work on Mac build.
This expands the "desc" command to list includes, defined, and flags. It
will now optionally annotate the source of the configs producting this list.
I folded the old "tree" and "deps" subcommands into the new expanded desc
command.

This works on the Mac build. Some more Mac flags are set in the environment
and I added Mac extensions to some of the lists. It does not yet completely
work, however.

This adds a lot of Mac/GCC stuff to the current build. I tracked back the
current set of defines that the GYP build sets, and set them in the GN build.

BUG=
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/21636002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8338eeabe776455a934ccf099cac7bd87fa57588