Adjust histograms for CA's BR compliance to also consider the notBefore date
The histograms were only considering the expiration date, as that is
all that Appendix B lists as applicable for key sizes. However,
Appendix B is only relevant if the certificate is in scope of the BRs,
which itself is gated on issuance date (BRs 1.0 were adopted with an
effective date of 2012-01-07, although root programs were slower to
require audit compliance).

BUG=102949
R=wtc,asvitkine

Review URL: https://chromiumcodereview.appspot.com/22606007

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d70e808c7e3ebcae52143509d14ab37eaefe2b0c