[Telemetry] Fix page cycler reliability on mac.
The page cyclers tend to fail on mac due to hitting the maximum number of
open file descriptor limit. This was the reason typical 25 is disabled, the
reason several page cyclers are failing on the waterfall now, and the reason
for the connection closing hack.

This patch raises the max fd limit rather than trying to stay below the limit.
It removes the old hack, and re-enables typical 25.

This problem is very hard to reproduce locally, so I'm not 100% confident in
this patch. I'd like to land it, keep and eye on it, and revert if the failures
don't get significantly better.

BUG=269727,268646

Review URL: https://chromiumcodereview.appspot.com/23273002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c947202f15f1df3d1cc2224c10a4d981cd35e017