Remove nss hack in chrome.
This does more correct dependencies and flags for linking NSS into Chrome.

It adds buildfiles for skia. The library compiles on Linux but likely not other systems at this point and is still a bit of a mess.

Defines are now escaped properly so you can have quotes in defined values (this was needed for Skia).

This adds a describe command for printing out the applying ld flags, and removes duplicates earlier using a new OrderedSet container.

BUG=
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/23480042

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b1e468fff539bd050302b9154db2aee8bad1a696