Initial WebUI for DOM Distiller.
The DOM Distiller component will contain code for an experimental
prototype for distilling the core part of a web page.

To enable this feature, use the command line flag --enable-distiller.

The webui/ folder depends on content, but given iOS at the time being
supports the usage of WebUI, it is kept a a top level folder in the
component instead of in content/.

BUG=288015

Review URL: https://chromiumcodereview.appspot.com/23503042

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bea0cd590e37917a3252b4a54bd07178130862e5