Initial submit of ChromeDebug extension.
This is a re-upload of a previous patch due to some cs files 
breaking git's copy/rename detection, so a new issue had to be 
created and uploaded with --no-find-copies.

New to this upload is the addition of a LICENSE file (which 
is displayed when you try to install the extension), as well 
as copyright headers in source files.

BUG=0

Review URL: https://chromiumcodereview.appspot.com/23480085

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a5bee464b29124393a84c46216c21439a9e9e13f