[password generation] Upload possible account creation forms via Autofill
We consider that a form may be used for account creation if the first time
the saved credential is used it is on a different form, and the original form
has at least 3 text input fields.

This change does not enable uploading for GNOME or KWallet users.

BUG=240559

Review URL: https://chromiumcodereview.appspot.com/23580005

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ecf8fa991faee69cdc29f216cb276918a81d2429