Allow MEDIA_KEYERR_OUTPUT to show up in histograms
The Widevine CDM may report a MEDIA_KEYERR_OUTPUT due to the output device
not having the appropriate protection. Allow the error to propogate up to
Chrome so it can be reported to the user and shown in UMA stats.

BUG=294501

Review URL: https://chromiumcodereview.appspot.com/23818007

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7aface6ec119ce31aa74d0b311ddf8f4ade618f6