Add instruction to install Visual Studio SDK for ChromeDebug.
For C# projects, if a ProjectType GUID is not available, the vague error
"The project type is not supported by this installation" is given.
In this case the issue is usually that the Visual Studio SDK is not
installed.

R=zturner@chromium.org

Review URL: https://codereview.chromium.org/26460003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 27909fa8cd4b89c8fff94348771d0ec56e4ca374