Docserver: Pull knowledge of host file systems into a single HostFileSystemProvider class, basically a rename of HostFileSystemCreator with a stricter notion of trunk vs branches.
In some ways this is a follow-up to r218556, except solving the problem in a
type-safe (rather than special-case) way, as requested in that review. I also
need it to be able to support "host" file systems as a content provider, see
bug 293678.

This patch also enforces that the "branch" of a channel is a string, and fixes a
bunch of callers, because a bug there was exposed as part of this work.

BUG= 273105, 293678
R=jyasskin@chromium.org

Review URL: https://codereview.chromium.org/26418002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c7ec92eb9f883847d956f81934099a11a4fce97b