Moved ignoring of --gtest_output argument to apply to all Telemetry tests.
This is needed for all tests to be able to upload their results to the
flakiness dashboard. I forgot to do this for the newly-added
context_lost GPU test, but it's better to just do it for all tests.

BUG=308675
R=dtu@chromium.org

Review URL: https://codereview.chromium.org/28073002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d4748634795e820e53a82c61c7ffd15d5ad3040e